Use features of active record to reduce queries
The features used here are includes, inverse_of, and pluck.
Includes is used to prevent 1+N queries. In this case each DataSet was making a separate query its user.
Inverse_of helps ActiveRecord reuse existing objects. So in this case the project object will be the same as data_set.project.
Pluck is an efficient way to return just the value of one field.